+++ Cargo.toml
@@ -3,11 +3,15 @@ name = "clip-mash"
 version = "0.7.2"
 edition = "2021"
 
+[workspace]
+members = ["clip-mash-types"]
+
 [dependencies]
 aubio = { version = "0.2.1", features = ["bindgen"] }
 axum = { version = "0.6.18", features = ["macros", "multipart"] }
 camino = { version =  "1.1.4", features = ["serde1"] }
 color-eyre = "0.6.2"
+clip-mash-types = { path = "clip-mash-types" }
 directories = "5.0.1"
 futures = "0.3.28"
 graphql_client = "0.12.0"
@@ -54,3 +58,4 @@ tracing-test = "0.2.4"
 
 [build-dependencies]
 typescript-type-def = "0.5.6"
+clip-mash-types = { path = "clip-mash-types" }
