+++ src/server/dtos.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 
 use camino::Utf8Path;
+use clip_mash_types::*;
 use serde::{Deserialize, Serialize};
 
 use crate::{
@@ -8,47 +9,9 @@ use crate::{
         database::{DbMarker, DbVideo, LocalVideoWithMarkers},
         stash_api::{find_scenes_query::FindScenesQueryFindScenesScenes, StashMarker},
     },
-    service::{clip::ClipOrder, generator::VideoResolution, Clip, MarkerId, Video, VideoId},
+    service::Video,
     util::{add_api_key, expect_file_name},
 };
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct TagDto {
-    pub name: String,
-    pub id: String,
-    pub marker_count: i64,
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct PerformerDto {
-    pub id: String,
-    pub scene_count: i64,
-    pub name: String,
-    pub image_url: Option<String>,
-    pub tags: Vec<String>,
-    pub rating: Option<i64>,
-    pub favorite: bool,
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct MarkerDto {
-    pub id: MarkerId,
-    pub video_id: VideoId,
-    pub primary_tag: String,
-    pub stream_url: String,
-    pub start: f64,
-    pub end: f64,
-    pub scene_title: Option<String>,
-    pub performers: Vec<String>,
-    pub file_name: Option<String>,
-    pub scene_interactive: bool,
-    pub tags: Vec<String>,
-    pub screenshot_url: Option<String>,
-    pub index_within_video: usize,
-}
 
 impl From<StashMarker> for MarkerDto {
     fn from(value: StashMarker) -> Self {
@@ -90,16 +53,6 @@ impl From<DbMarker> for MarkerDto {
             video_id: VideoId::LocalFile(value.video_id),
         }
     }
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct VideoDto {
-    pub id: VideoId,
-    pub title: String,
-    pub performers: Vec<String>,
-    pub file_name: String,
-    pub interactive: bool,
 }
 
 impl From<FindScenesQueryFindScenesScenes> for VideoDto {
@@ -148,69 +101,6 @@ impl From<Video> for VideoDto {
     }
 }
 
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct SelectedMarker {
-    pub id: MarkerId,
-    pub video_id: VideoId,
-    pub selected_range: (f64, f64),
-    pub index_within_video: usize,
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase", tag = "type")]
-pub struct RandomizedClipOptions {
-    pub base_duration: f64,
-    pub divisors: Vec<f64>,
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase", tag = "type")]
-pub enum PmvClipOptions {
-    Randomized(RandomizedClipOptions),
-    Songs {
-        beats_per_measure: usize,
-        // TODO allow randomizing
-        cut_after_measure_count: usize,
-    },
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase", tag = "type")]
-pub enum ClipOptions {
-    Pmv {
-        song_ids: Vec<i64>,
-        clips: PmvClipOptions,
-    },
-    Default(RandomizedClipOptions),
-}
-
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct CreateClipsBody {
-    pub clip_order: ClipOrder,
-    pub split_clips: bool,
-    pub markers: Vec<SelectedMarker>,
-    pub seed: Option<String>,
-    pub clips: ClipOptions,
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ClipsResponse {
-    pub clips: Vec<Clip>,
-    pub streams: HashMap<String, String>,
-    pub videos: Vec<VideoDto>,
-    pub beat_offsets: Option<Vec<f32>>,
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ListVideoDto {
-    pub video: VideoDto,
-    pub markers: Vec<MarkerDto>,
-}
-
 impl From<LocalVideoWithMarkers> for ListVideoDto {
     fn from(value: LocalVideoWithMarkers) -> Self {
         ListVideoDto {
@@ -219,57 +109,3 @@ impl From<LocalVideoWithMarkers> for ListVideoDto {
         }
     }
 }
-
-#[derive(Deserialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct CreateVideoBody {
-    pub file_name: String,
-    pub clips: Vec<Clip>,
-    pub selected_markers: Vec<SelectedMarker>,
-    pub output_resolution: VideoResolution,
-    pub output_fps: u32,
-    pub song_ids: Vec<i64>,
-    pub music_volume: Option<f64>,
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct StashScene {
-    pub id: String,
-    pub performers: Vec<String>,
-    pub image_url: Option<String>,
-    pub title: String,
-    pub studio: Option<String>,
-    pub tags: Vec<String>,
-    pub rating: Option<i64>,
-    pub interactive: bool,
-    pub marker_count: usize,
-}
-
-impl StashScene {
-    pub fn from(scene: FindScenesQueryFindScenesScenes, api_key: &str) -> Self {
-        StashScene {
-            id: scene.id,
-            performers: scene.performers.into_iter().map(|p| p.name).collect(),
-            image_url: scene.paths.screenshot.map(|url| add_api_key(&url, api_key)),
-            title: scene.title.unwrap_or_default(),
-            studio: scene.studio.map(|s| s.name),
-            tags: scene.tags.into_iter().map(|t| t.name).collect(),
-            rating: scene.rating100,
-            interactive: scene.interactive,
-            marker_count: scene.scene_markers.len(),
-        }
-    }
-}
-
-pub type Api = (
-    StashScene,
-    CreateVideoBody,
-    CreateClipsBody,
-    ListVideoDto,
-    ClipsResponse,
-    VideoDto,
-    MarkerDto,
-    PerformerDto,
-    TagDto,
-);
